refactor: optimize backend pages performance
- Move sections array outside component in dashboard page to prevent recreation on each render
- Fix incorrect imports in deliveryMethods and reserves components (use ~/trpc/react)
- Memoize CSV parsing operations in import-alumni page to avoid repeated calculations
- Improve overall rendering performance and reduce unnecessary computations
